refactoring - remove unnecessary config file
